+++ src/condition/type/Number.js
@@ -3,8 +3,6 @@ var util = require('util');
 var Condition = require('../Condition'),
     errorMessage = require('../../messages').Type;
 
-var coercionRegex = /^-?\d+(\.\d+)?$/;
-
 /**
  * @class lewd.condition.type.Number
  * @extends {lewd.condition.Condition}
@@ -24,10 +22,6 @@ util.inherits(NumberTypeCondition, Condition);
 NumberTypeCondition.prototype.validate = function (value, path) {
     if (typeof value === 'number' && isFinite(value)) {
         return value;
-    }
-    
-    if (this.coerce && typeof value === 'string' && coercionRegex.test(value)) {
-        return parseFloat(value);
     }
 
     this.reject(value, path, errorMessage, { type: 'number' });